requirements.txt
fastapi
uvicorn
django
psycopg2-binary
python-dotenv
pydantic[email]
